support-app-backend/requirements.txt
fastapi[standard]==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
asyncpg>=0.30.0,<0.31.0
psycopg[binary]>=3.2.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.13,<0.4.0
langchain-google-vertexai>=2.0.0
langchain-postgres>=0.0.16
langgraph>=0.2.0
pgvector>=0.2.5,<0.4.0
google-cloud-aiplatform>=1.75.0
google-auth>=2.0.0
google-adk>=0.2.0
# Document processing libraries
Pillow>=12.1.0
pytesseract>=0.3.13
openpyxl>=3.1.5
python-docx>=1.2.0
PyPDF2>=3.0.1
tesseract>=0.1.3
aiohttp>=3.13.3
slowapi>=0.1.9

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0

# Caching
redis>=5.0.0

# Database migrations
alembic>=1.14.0
